requirements.txt
Flask==3.0.0
requests==2.31.0
python-dotenv==1.1.0
python-weather==2.0.7
termcolor==3.0.1
smartfunc==0.2.0
